Add stub module install
This commit adds a new level to the manager. Now the top two subcommands
are 'env' and 'mod', for installing morloc environments and modules. All
installed modules will go into the library for the currently selected morloc
version.

Modules may be imported from local files, remotes (github, bitbucket, or
gitlab, currently), or from urls.

The import syntax is
  [source:]<module>[@[reference_type:]reference]

The source may be "file", "http", "https", "github", etc.

The module may be a single module name (for core libraries), a local
path, a username/repo combo for remote repositories, or a URL.

The reference may be a version, hash, or branch name. These three types
may be inferred from the reference string or specified explicitly with
the reference_type label.

Here are a few examples:

 * install the latest `root` module from the standard library
     morloc mod install root

 * install a specified version of the `root` module from the standard library
     morloc mod install root@abcdef1234

 * specify that abcdef1234 is a hash (not a branch name)
     morloc mod install root@hash:abcdef1234

 * install a local module from the working directory
     morloc mod install .

 * install a local module from a path
     morloc mod install ~/src/momod

 * install the dev version of a third party module from github
     morloc mod install github:username/repo@branch:dev

 * install a module from a URL
     morloc mod install https:path/to/my/module.tar.gz@hash:5457644959cf712a691380075eaa3cea
     - here the hash is the expected md5sum of the file, for security